actualizacion general 2

+++ database/migrations/2024_11_08_133811_create_servicios_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('codigo')->unique();
             $table->string('nombre');
-            $table->float('valor_hora')->nullable();
+            $table->decimal('valor_hora', 20, 2)->nullable();
             $table->timestamps();
         });
     }
